refactor: split bloated ib-client monolith into modular submodules
